requirements.txt
flask
azure-storage-blob
psycopg2-binary
python-dotenv
flask-cors